requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
streamlit==1.28.0
python-multipart==0.0.9
pydantic==2.10.6
pydantic-settings==2.5.2
sqlalchemy==2.0.23
aiosqlite==0.19.0
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.3

# LangChain dependencies
langchain==0.3.0
langchain-community==0.3.0
langchain-dashscope>=0.1.0

# Vector database
faiss-cpu==1.8.0

# AI/ML
dashscope==1.14.0
numpy==1.26.2

# Document processing
pypdf>=5.0.0
python-docx==1.1.0
openpyxl==3.1.2
reportlab==4.0.7
pdf2image==1.16.3

# Data and visualization
pandas==2.1.3
plotly==5.18.0
pillow==10.1.0

# Bot platforms
python-telegram-bot==20.7
botbuilder-core==4.15.0
botbuilder-schema==4.15.0
botbuilder-integration-aiohttp==4.15.0
